+++ src/texture.c
@@ -103,7 +103,7 @@ texture *add_uv_texture(memory_region *region) {
 }
 
 // Checker Texture -- takes two textures and returns them in a checker pattern 
-checker_texture checker_texture_new(void *odd, void *even, float freq) {
+checker_texture checker_texture_new(void *odd, void *even, double freq) {
 	checker_texture txt = {};
 	txt.odd = odd;
 	txt.even = even;
@@ -111,7 +111,7 @@ checker_texture checker_texture_new(void *odd, void *even, float freq) {
 	return txt;
 }
 
-texture make_checker_texture(texture *odd, texture *even, float freq) {
+texture make_checker_texture(texture *odd, texture *even, double freq) {
 	texture txt = {};
 	txt.TextureColor = (*CheckerTextureColor);
 	checker_texture checkertxt = checker_texture_new(odd, even, freq);
@@ -120,7 +120,7 @@ texture make_checker_texture(texture *odd, texture *even, float freq) {
 	return txt;
 }
 
-texture *add_checker_texture(memory_region *region, texture *odd, texture *even, float freq) {
+texture *add_checker_texture(memory_region *region, texture *odd, texture *even, double freq) {
 	texture checkertxt = make_checker_texture(odd, even, freq);
 	texture *checkerptr = (texture *)memory_region_add(region, &checkertxt, sizeof(texture));
 	return checkerptr;
@@ -130,7 +130,7 @@ texture *add_checker_texture(memory_region *region, texture *odd, texture *even,
 // this allows flat shapes to have tiles
 // Since it's identical in every way, this will just use checker_texture_new
 
-texture make_uv_checker_texture(texture *odd, texture *even, float freq) {
+texture make_uv_checker_texture(texture *odd, texture *even, double freq) {
 	texture txt = {};
 	txt.TextureColor = (*UVCheckerTextureColor);
 	checker_texture checkertxt = checker_texture_new(odd, even, freq);
@@ -139,14 +139,14 @@ texture make_uv_checker_texture(texture *odd, texture *even, float freq) {
 	return txt;
 }
 
-texture *add_uv_checker_texture(memory_region *region, texture *odd, texture *even, float freq) {
+texture *add_uv_checker_texture(memory_region *region, texture *odd, texture *even, double freq) {
 	texture checkertxt = make_uv_checker_texture(odd, even, freq);
 	texture *checkerptr = (texture *)memory_region_add(region, &checkertxt, sizeof(texture));
 	return checkerptr;
 }
 
 // noise Noise Texture
-noise_texture noise_texture_new(memory_region *region, noise *noise, float scale, int pointcount, void *texture) {
+noise_texture noise_texture_new(memory_region *region, noise *noise, double scale, int pointcount, void *texture) {
 	noise_texture p = {};
 	p.noise = noise;
 	p.texture = texture;
@@ -154,7 +154,7 @@ noise_texture noise_texture_new(memory_region *region, noise *noise, float scale
 	return p;
 }
 
-texture make_noise_texture(memory_region *region, noise *noise, float scale, int pointcount, texture *text) {
+texture make_noise_texture(memory_region *region, noise *noise, double scale, int pointcount, texture *text) {
 	texture txt = {};
 	txt.TextureColor = (*NoiseTextureColor);
 	noise_texture noisetxt = noise_texture_new(region, noise, scale, pointcount, text);
@@ -163,14 +163,14 @@ texture make_noise_texture(memory_region *region, noise *noise, float scale, int
 	return txt;
 }
 
-texture *add_noise_texture(memory_region *region, noise *noise, float scale) {
+texture *add_noise_texture(memory_region *region, noise *noise, double scale) {
 	// NOTE : this is not good practice imo, but for convencience sake
 	texture noisetext = make_noise_texture(region, noise, scale, 256, NULL);
 	texture *noiseptr = (texture *)memory_region_add(region, &noisetext, sizeof(texture));
 	return noiseptr;
 }
 
-texture *add_colored_noise_texture(memory_region *region, noise *noise, float scale, texture *text) {
+texture *add_colored_noise_texture(memory_region *region, noise *noise, double scale, texture *text) {
 	// NOTE : this is not good practice imo, but for convencience sake
 	texture noisetext = make_noise_texture(region, noise, scale, 256, text);
 	texture *noiseptr = (texture *)memory_region_add(region, &noisetext, sizeof(texture));
@@ -182,7 +182,7 @@ texture *add_colored_noise_texture(memory_region *region, noise *noise, float sc
 // Pass bits (must be in [1 - 10]) and the internal noise structure will 
 // use that many points.
 // It basically determines how often the pattern repeats.
-texture *add_noise_texture_sized(memory_region *region, float scale, int bits) {
+texture *add_noise_texture_sized(memory_region *region, double scale, int bits) {
 	// NOTE : this is not good practice imo, but for convencience sake
 	int pointcount = (bits > 10 || bits <= 0) ? 256 : pow(2, bits);
 	texture noisetext = make_noise_texture(region, NULL, scale, pointcount, NULL);
@@ -190,7 +190,7 @@ texture *add_noise_texture_sized(memory_region *region, float scale, int bits) {
 	return noiseptr;
 }
 
-texture *add_colored_noise_texture_sized(memory_region *region, float scale, int bits, texture *text) {
+texture *add_colored_noise_texture_sized(memory_region *region, double scale, int bits, texture *text) {
 	// NOTE : this is not good practice imo, but for convencience sake
 	int pointcount = (bits > 10 || bits <= 0) ? 256 : pow(2, bits);
 	texture noisetext = make_noise_texture(region, NULL, scale, pointcount, text);
@@ -199,7 +199,7 @@ texture *add_colored_noise_texture_sized(memory_region *region, float scale, int
 }
 
 // noise Turbulence -- same as noise_noise_texture but calls noise_turbulence instead of noise_noise
-texture *add_noise_turbulence_texture(memory_region *region, noise *noise, float scale) {
+texture *add_noise_turbulence_texture(memory_region *region, noise *noise, double scale) {
 	texture noisetext = make_noise_texture(region, noise, scale, 256, NULL);
 	noisetext.TextureColor = (*FBMTextureColor);
 	noisetext.id = FBM;
@@ -207,7 +207,7 @@ texture *add_noise_turbulence_texture(memory_region *region, noise *noise, float
 	return noiseptr;
 }
 
-texture *add_colored_noise_turbulence_texture(memory_region *region, noise *noise, float scale, texture *text) {
+texture *add_colored_noise_turbulence_texture(memory_region *region, noise *noise, double scale, texture *text) {
 	texture noisetext = make_noise_texture(region, noise, scale, 256, text);
 	noisetext.TextureColor = (*FBMTextureColor);
 	noisetext.id = FBM;
@@ -215,7 +215,7 @@ texture *add_colored_noise_turbulence_texture(memory_region *region, noise *nois
 	return noiseptr;
 }
 
-texture *add_marbled_noise_texture(memory_region *region, noise *noise, float scale, texture *text) {
+texture *add_marbled_noise_texture(memory_region *region, noise *noise, double scale, texture *text) {
 	texture noisetext = make_noise_texture(region, noise, scale, 256, text);
 	noisetext.TextureColor = (*MarbledNoiseTextureColor);
 	noisetext.id = MarbledNoise;
@@ -224,7 +224,7 @@ texture *add_marbled_noise_texture(memory_region *region, noise *noise, float sc
 }
 
 // gradient noise Textures -- noise textures that use two colors.
-gradient_noise_texture gradient_noise_texture_new(memory_region *region, noise *noise, float scale, int pointcount, void *colA, void *colB) {
+gradient_noise_texture gradient_noise_texture_new(memory_region *region, noise *noise, double scale, int pointcount, void *colA, void *colB) {
 	gradient_noise_texture p = {};
 	p.noise = noise;
 	p.colA = colA;
@@ -233,7 +233,7 @@ gradient_noise_texture gradient_noise_texture_new(memory_region *region, noise *
 	return p;
 }
 
-texture make_gradient_noise_texture(memory_region *region, noise *noise, float scale, int pointcount, texture *colA, texture *colB) {
+texture make_gradient_noise_texture(memory_region *region, noise *noise, double scale, int pointcount, texture *colA, texture *colB) {
 	texture txt = {};
 	txt.TextureColor = (*UNDEFINED_TextureColor);
 	gradient_noise_texture noisetxt = gradient_noise_texture_new(region, noise, scale, pointcount, colA, colB);
@@ -242,7 +242,7 @@ texture make_gradient_noise_texture(memory_region *region, noise *noise, float s
 	return txt;
 }
 
-texture *add_noise_sincos_texture(memory_region *region, noise *noise, float scale, texture *colA, texture *colB) {
+texture *add_noise_sincos_texture(memory_region *region, noise *noise, double scale, texture *colA, texture *colB) {
 	int pointcount = 256;
 	texture mp = make_gradient_noise_texture(region, noise, scale, pointcount, colA, colB);
 	mp.TextureColor = (*NoiseSinCosTextureColor);
@@ -253,7 +253,7 @@ texture *add_noise_sincos_texture(memory_region *region, noise *noise, float sca
 
 
 // FBM Modifier -- take some other texture and return the value when you mess with the point
-fbm_texture fbm_modifier_new(memory_region *region, noise *noise, float hurst, int octaves, void *text) {
+fbm_texture fbm_modifier_new(memory_region *region, noise *noise, double hurst, int octaves, void *text) {
 	fbm_texture fbm_mod = {};
 	fbm_mod.hurst = hurst;
 	fbm_mod.octaves = octaves;
@@ -262,7 +262,7 @@ fbm_texture fbm_modifier_new(memory_region *region, noise *noise, float hurst, i
 	return fbm_mod;
 }
 
-texture make_fbm_modifier(memory_region *region, noise *noise, float hurst, int octaves, texture *text) {
+texture make_fbm_modifier(memory_region *region, noise *noise, double hurst, int octaves, texture *text) {
 	texture txt = {};
 	txt.TextureColor = (*FBMModifierTextureColor);
 	fbm_texture fbm_mod = fbm_modifier_new(region, noise, hurst, octaves, text);
@@ -270,7 +270,7 @@ texture make_fbm_modifier(memory_region *region, noise *noise, float hurst, int 
 	txt.id = FBM;
 	return txt;
 }
-texture *add_fbm_modifier(memory_region *region, noise *noise, float hurst, int octaves, texture *text) {
+texture *add_fbm_modifier(memory_region *region, noise *noise, double hurst, int octaves, texture *text) {
 	texture fbmtxt = make_fbm_modifier(region, noise, hurst, octaves, text);
 	texture *fbmptr = (texture *)memory_region_add(region, &fbmtxt, sizeof(texture));
 	return fbmptr;
@@ -290,7 +290,7 @@ texture *add_level_curve_texture(memory_region *region, vec3 intervals, vec3 wid
 	return (texture *)memory_region_add(region, &lct, sizeof(texture));
 }
 
-texture make_distance_texture(point3 point, float cutoff, texture *textA, texture *textB) {
+texture make_distance_texture(point3 point, double cutoff, texture *textA, texture *textB) {
 	texture dtxt = {};
 	dtxt.TextureColor = (*DistanceTextureColor);
 	dtxt.type.distance = distance_texture_new(point, fabs(cutoff), textA, textB);
@@ -298,7 +298,7 @@ texture make_distance_texture(point3 point, float cutoff, texture *textA, textur
 	return dtxt;
 }
 
-texture *add_distance_texture(memory_region *region, point3 point, float cutoff, texture *textA, texture *textB) {
+texture *add_distance_texture(memory_region *region, point3 point, double cutoff, texture *textA, texture *textB) {
 	texture dtxt = make_distance_texture(point, cutoff, textA, textB);
 	return (texture *)memory_region_add(region, &dtxt, sizeof(texture));
 }
@@ -306,111 +306,111 @@ texture *add_distance_texture(memory_region *region, point3 point, float cutoff,
 /* TextureColor 'Methods' */
 
 // Assume all normals are unit vectors already. Pretty sure they are.
-fcolor NormalTextureColor(void *self, float u, float v, vec3 pt, vec3 *normal) {
+fcolor NormalTextureColor(void *self, double u, double v, vec3 pt, vec3 *normal) {
 	fcolor norm_color = fcolor_new(fabs(normal->x), fabs(normal->y), fabs(normal->z));
 	return norm_color;
 }
 
-fcolor SignedNormalTextureColor(void *self, float u, float v, vec3 pt, vec3 *normal) {
-	fcolor norm_color = fcolor_new((normal->x < 0.0f) ? 0.0f : normal->x, (normal->y < 0.0f) ? 0.0f : normal->y, (normal->z < 0.0f) ? 0.0f : normal->z);
+fcolor SignedNormalTextureColor(void *self, double u, double v, vec3 pt, vec3 *normal) {
+	fcolor norm_color = fcolor_new((normal->x < 0.0) ? 0.0 : normal->x, (normal->y < 0.0) ? 0.0 : normal->y, (normal->z < 0.0) ? 0.0 : normal->z);
 	return norm_color;
 }
 
-fcolor ColorTextureColor(void *self, float u, float v, vec3 pt, vec3 *normal) {
+fcolor ColorTextureColor(void *self, double u, double v, vec3 pt, vec3 *normal) {
 	color_texture *txt = &((texture *)self)->type.color; 
 	return txt->rgb;
 }
 
-fcolor ImageTextureColor(void *self, float u, float v, point3 pt, vec3 *normal) {
+fcolor ImageTextureColor(void *self, double u, double v, point3 pt, vec3 *normal) {
 	image_texture *imgtxt = &((texture *)self)->type.image;
 	if (!imgtxt->pixels) return COLOR_UNDEFPURP;
-	int x = clamp(u, 0.0f, 1.0f) * (float)imgtxt->width;
-	int y = imgtxt->height - clamp(v, 0.0f, 1.0f) * (float)imgtxt->height;
+	int x = clamp(u, 0.0, 1.0) * (double)imgtxt->width;
+	int y = imgtxt->height - clamp(v, 0.0, 1.0) * (double)imgtxt->height;
 	if(x >= imgtxt->width) x = imgtxt->width - 1;
 	if(y >= imgtxt->height) y = imgtxt->height - 1;
 
 	unsigned char *pixel = imgtxt->pixels + (y * imgtxt->pitch) + (x * imgtxt->bytes_per_pixel);
-	fcolor result = fcolor_new((float)*pixel / 255.0f, (float)*(pixel + 1) / 255.0f, (float)*(pixel + 2) / 255.0f);
+	fcolor result = fcolor_new((double)*pixel / 255.0, (double)*(pixel + 1) / 255.0, (double)*(pixel + 2) / 255.0);
 	return result;
 }
 
-fcolor CheckerTextureColor(void *self, float u, float v, vec3 pt, vec3 *normal) {
+fcolor CheckerTextureColor(void *self, double u, double v, vec3 pt, vec3 *normal) {
 	checker_texture *chtxt = &((texture *)self)->type.checker;
-	float sines = sinf(chtxt->freq * pt.x) * sinf(chtxt->freq * pt.y) * sinf(chtxt->freq * pt.z);
-	if (sines < 0.0f)
+	double sines = sinf(chtxt->freq * pt.x) * sinf(chtxt->freq * pt.y) * sinf(chtxt->freq * pt.z);
+	if (sines < 0.0)
 		return ((texture *)chtxt->odd)->TextureColor(chtxt->odd, u, v, pt, normal);
 	return ((texture *)chtxt->even)->TextureColor(chtxt->even, u, v, pt, normal);
 }
 
-fcolor UVCheckerTextureColor(void *self, float u, float v, vec3 pt, vec3 *normal) {
+fcolor UVCheckerTextureColor(void *self, double u, double v, vec3 pt, vec3 *normal) {
 	checker_texture *chtxt = &((texture *)self)->type.checker;
-	const float twopi = 2.0f * pi;
-	float sines = sinf(chtxt->freq * u * twopi) * sinf(chtxt->freq * v * twopi);
-	if (sines < 0.0f)
+	const double twopi = 2.0 * pi;
+	double sines = sinf(chtxt->freq * u * twopi) * sinf(chtxt->freq * v * twopi);
+	if (sines < 0.0)
 		return ((texture *)chtxt->odd)->TextureColor(chtxt->odd, u, v, pt, normal);
 	return ((texture *)chtxt->even)->TextureColor(chtxt->even, u, v, pt, normal);
 }
 
-fcolor NoiseTextureColor(void *self, float u, float v, vec3 pt, vec3 *normal) {
+fcolor NoiseTextureColor(void *self, double u, double v, vec3 pt, vec3 *normal) {
 	noise_texture *noise = &((texture *)self)->type.noise;
 	fcolor col;
 	if (noise->texture != NULL) {
 		col = ((texture *)noise->texture)->TextureColor(noise->texture, u, v, pt, normal);
 	} else {
-		col = fcolor_new(1.0f, 1.0f, 1.0f);
+		col = fcolor_new(1.0, 1.0, 1.0);
 	}
 	point3 scaled_pt = vec3_mul(pt, noise->scale);
-	float n = (1.0f + noise->noise->Noise(noise->noise->source, &scaled_pt)) * 0.5f;
+	double n = (1.0 + noise->noise->Noise(noise->noise->source, &scaled_pt)) * 0.5;
 	return color_mul(col, n);
 }
 
-fcolor FBMTextureColor(void *self, float u, float v, vec3 pt, vec3 *normal) {
+fcolor FBMTextureColor(void *self, double u, double v, vec3 pt, vec3 *normal) {
 	noise_texture *noise = &((texture *)self)->type.noise;
 	int depth = 7;
 	fcolor col;
 	if (noise->texture != NULL) {
 		col = ((texture *)noise->texture)->TextureColor(noise->texture, u, v, pt, normal);
 	} else {
-		col = fcolor_new(1.0f, 1.0f, 1.0f);
+		col = fcolor_new(1.0, 1.0, 1.0);
 	}
 
 	point3 scaled_pt = vec3_mul(pt, noise->scale);
-	return color_mul(col, fbm(noise->noise, pt, 0.5f, depth));
+	return color_mul(col, fbm(noise->noise, pt, 0.5, depth));
 }
 
-fcolor MarbledNoiseTextureColor(void *self, float u, float v, vec3 pt, vec3 *normal) {
+fcolor MarbledNoiseTextureColor(void *self, double u, double v, vec3 pt, vec3 *normal) {
 	noise_texture *noise = &((texture *)self)->type.noise;
 	int depth = 7;
 	fcolor col;
 	if (noise->texture != NULL) {
 		col = ((texture *)noise->texture)->TextureColor(noise->texture, u, v, pt, normal);
 	} else {
-		col = fcolor_new(1.0f, 1.0f, 1.0f);
+		col = fcolor_new(1.0, 1.0, 1.0);
 	}
-	float n = 0.5f * (1.0f + sinf(pt.y + 10.0f * fabsf(fbm(noise->noise, pt, 0.5f, depth))));
+	double n = 0.5 * (1.0 + sinf(pt.y + 10.0 * fabsf(fbm(noise->noise, pt, 0.5, depth))));
 	return color_mul(col, n);
 }
 
-fcolor NoiseSinCosTextureColor(void *self, float u, float v, vec3 pt, vec3 *normal) {
+fcolor NoiseSinCosTextureColor(void *self, double u, double v, vec3 pt, vec3 *normal) {
 	gradient_noise_texture *noise = &((texture *)self)->type.gradient_noise;
 	int depth = 7;
 	vec3 swizzled = vec3_new(pt.z, pt.x, pt.y);
-	float turb = sinf(noise->scale * 10.0f * fabsf(fbm(noise->noise, pt, 0.5f, depth)));
-	turb *= cosf(noise->scale * fabsf(fbm(noise->noise, swizzled, 0.5f, depth)));
-	if (turb > 0.0f) {
+	double turb = sinf(noise->scale * 10.0 * fabsf(fbm(noise->noise, pt, 0.5, depth)));
+	turb *= cosf(noise->scale * fabsf(fbm(noise->noise, swizzled, 0.5, depth)));
+	if (turb > 0.0) {
 		return color_mul(((texture *)noise->colA)->TextureColor(noise->colA, u, v, pt, normal), turb);
 	} else {
 		return color_mul(((texture *)noise->colB)->TextureColor(noise->colB, u, v, pt, normal), fabs(turb));
 	}
 }
 
-fcolor FBMModifierTextureColor(void *self, float u, float v, vec3 pt, vec3 *normal) {
+fcolor FBMModifierTextureColor(void *self, double u, double v, vec3 pt, vec3 *normal) {
 	fbm_texture *fbm_mod = &((texture *)self)->type.fbm;
 	vec3 warped = vec3_mul(pt, fbm(fbm_mod->noise, pt, fbm_mod->hurst, fbm_mod->octaves));
 	return ((texture *)fbm_mod->text)->TextureColor(fbm_mod->text, u, v, warped, normal);
 }
 
-fcolor UNDEFINED_TextureColor(void *self, float u, float v, point3 pt, vec3 *normal) {
+fcolor UNDEFINED_TextureColor(void *self, double u, double v, point3 pt, vec3 *normal) {
 	return COLOR_UNDEFPURP;
 }
 
@@ -423,7 +423,7 @@ vec3 vec3_fmodf(vec3 *a, vec3 *m) {
 	return r;
 }
 
-fcolor LevelCurveTextureColor(void *self, float u, float v, vec3 pt, vec3 *normal) {
+fcolor LevelCurveTextureColor(void *self, double u, double v, vec3 pt, vec3 *normal) {
 	level_curve_texture lc_txt = ((texture *)self)->type.level_curve;
 	vec3 abs_pt = vec3_abs(pt);
 	vec3 r = vec3_fmodf(&abs_pt, &lc_txt.intervals);
@@ -431,10 +431,10 @@ fcolor LevelCurveTextureColor(void *self, float u, float v, vec3 pt, vec3 *norma
 	return ((texture *)lc_txt.defaultColor)->TextureColor(((texture *)lc_txt.defaultColor), u, v, pt, normal);
 }
 
-fcolor DistanceTextureColor(void *self, float u, float v, vec3 pt, vec3 *normal) {
+fcolor DistanceTextureColor(void *self, double u, double v, vec3 pt, vec3 *normal) {
 	distance_texture dtxt = ((texture *)self)->type.distance;
 	fcolor colA = (dtxt.textA) ? ((texture *)dtxt.textA)->TextureColor(dtxt.textA, u, v, pt, normal) : COLOR_WHITE;
-	float dist = vec3_len(vec3_sub(pt, dtxt.pt));
+	double dist = vec3_len(vec3_sub(pt, dtxt.pt));
 	if (dist >= dtxt.cutoff) return colA;
 	fcolor colB = (dtxt.textB) ? ((texture *)dtxt.textB)->TextureColor(dtxt.textB, u, v, pt, normal) : COLOR_BLACK;
